weighted the gacha system.
Dice is now the weight of the brush (Higher = more likely). Not certain how to explain. all weights are added together and then a random one is selected, but the weight sizes are based on the weight. a weight of 1 and a weight of 7 has the 7 weight be activated 7/8 times.